Add people to group

+++ app/Http/Controllers/GroupController.php
@@ -8,6 +8,7 @@ use App\Models\Group;
 use Illuminate\Http\Request;
 use App\Models\GroupParticipant;
 use File;
+
 class GroupController extends Controller
 {
     public function createGroup(Request $request)
@@ -43,30 +44,60 @@ class GroupController extends Controller
         return response()->json($data);
     }
 
-    public function updateGroup( Request $request ){
-        $group = Group::findOrFail( $request->groupId  );
+    public function updateGroup(Request $request)
+    {
+        if ($request->ajax()) {
+            $group = Group::findOrFail($request->groupId);
 
-        $group->update(["name" => $request->name]);
-        return response()->json(['success' => 'Group update successfully!!']);
+            $group->update(["name" => $request->name]);
+            return response()->json(['success' => 'Group update successfully!!']);
+        }
     }
 
-    public function deleteGroup( Request $request ){
-        $chats = Chat::where('group_id', $request->groupId);
-        $chats->each(function ($chat) {
-            $fileName = $chat->file_name;
-            $path = storage_path("app/public/chat/{$fileName}");
+    public function deleteGroup(Request $request)
+    {
+        if ($request->ajax()) {
+            $chats = Chat::where('group_id', $request->groupId);
+            $chats->each(function ($chat) {
+                $fileName = $chat->file_name;
+                $path = storage_path("app/public/chat/{$fileName}");
 
-            if (File::exists($path)) {
-                File::delete($path);
+                if (File::exists($path)) {
+                    File::delete($path);
+                }
+            });
+            $chats->delete();
+
+            $group = Group::with('participants')->findOrFail($request->groupId);
+            $group->participants()->delete();
+            $group->delete();
+            return response()->json(['success' => 'Group deleted successfully!!']);
+        }
+    }
+
+    public function addPeopleToGroup(Request $request)
+    {
+        if ($request->ajax()) {
+            if (isset($request->user_id) && !empty($request->user_id)) {
+                $group_id = $request->groupId;
+                $userIds = explode(',', $request->user_id);
+
+                collect($userIds)->each(function ($userId) use ($group_id) {
+                    GroupParticipant::create([
+                        'user_id' => $userId,
+                        'group_id' => $group_id,
+                        'status' => 0,
+                    ]);
+                });
             }
-        });
-        $chats->delete();
 
-        $group = Group::with('participants')->findOrFail( $request->groupId  );
-        $group->participants()->delete();
-        $group->delete();
-        return response()->json(['success' => 'Group deleted successfully!!']);
+            $data = [
+                "success" => "People added successfuly to group!!",
+            ];
+            return response()->json($data);
+        }
     }
+
     public function loadSuggestedPeople(Request $request)
     {
         if ($request->ajax()) {
@@ -74,5 +105,4 @@ class GroupController extends Controller
             return view('e-learning.course.instructor.message-group.suggested-people', $data);
         }
     }
-
 }
